Add files via upload
The final project of digital image processing : here is the description of project:
Course Project 5-Hand
A picture of a hand is given as follows (image file name: hand.jpg).
① Convert the original image to HSV color space.
② Initially locate the hand area according to the threshold in the HSV 
color space.
③ Use morphological operations or preprocessing methods to denoise 
and remove holes in the positioned area.
④ Display the extracted hand picture in the program, and calculate the 
average RGB color value of the skin color area according to the pixel 
values in the extracted hand area, and output the result.
⑤ Tip: The HSV color range of the skin tone area is generally between 
[0, 20, 70] and [20, 255, 255]. You can also use your own picture to 
achieve the effect in the example. The code language used is python, 
and C++ is also acceptable.
Requirements: The course report needs to give thought analysis, 
algorithm steps, code screenshots, and display of processing results.